authentication and authorization

+++ backend/controllers/goalControllers/index.js
@@ -1,4 +1,5 @@
 const Goals = require('../../models/goalsModel')
+const Users = require('../../models/usersModel')
 
 module.exports = {
 
@@ -6,7 +7,7 @@ module.exports = {
     // @route   GET  /v1/api/goals
     // @acess   Private
     getGoals: async(req, res, next) => {
-        const goals = await Goals.find()
+        const goals = await Goals.find({ user: req.user.id })
 
         res.status(200).json({ message: 'Successfully get goals', items: goals })
     },
@@ -19,7 +20,10 @@ module.exports = {
             return res.status(400).json({ error: 'Please add some text field' })
         }
 
-        const createdGoal = await Goals.create({ text: req.body.text })
+        const createdGoal = await Goals.create({
+            user: req.user.id,
+            text: req.body.text
+        })
 
         res.status(201).json({ message: 'Sucessfully create goal', item: createdGoal })
     },
@@ -34,6 +38,16 @@ module.exports = {
 
         if (!goal) {
             return res.status(400).json({ error: 'Goal not found by ID' })
+        }
+
+        const user = await Users.findById(req.user.id)
+
+        if (!user) {
+            return res.status(400).json({ error: 'User not found please try to Sign In again!' })
+        }
+
+        if (goal.user.toString() !== user.id) {
+            return res.status(401).json({ error: 'User not authorized to update this goal' })
         }
 
         const updatedGoal = await Goals.findByIdAndUpdate(id, req.body, { new: true })
@@ -51,6 +65,16 @@ module.exports = {
 
         if (!goal) {
             return res.status(400).json({ error: 'Goal not found by ID' })
+        }
+
+        const user = await Users.findById(req.user.id)
+
+        if (!user) {
+            return res.status(400).json({ error: 'User not found' })
+        }
+
+        if (goal.user.toString() !== user.id) {
+            return res.status(401).json({ error: 'User not authorized to update this goal' })
         }
 
         const deletedGoal = await Goals.findByIdAndDelete(id)
